refactor: (#15) features/main의 도메인 로직을 entities (room, chat)로 분리
