+++ src/entities/room/constants/roomConstants.ts
