+++ src/features/main/model/index.ts
@@ -1,4 +1,3 @@
-export * from './types';
 export * from './useChat';
 export * from './useCreateRoomForm';
 export * from './useMain';
